example of how to access the dataframe-js module
This file is an example of how to use the dataframe-js module. For the purposes of this project it appears to be suitable substitute for the pandas library as used in the demo code from OpenRouteService. 
If used in a new folder, first initialize a package.json file with:
 npm init --yes
If used in folder or project that has already been initialized then skip the above step and just install with:
npm install datagram-js